\name{Y}
\alias{Y}
\docType{data}
\title{
Observation matrix Y
}
\description{
An example of observation matrix
}
\usage{data("Y")}
\format{
  The format is:
 num [1:50] 11 8 3 3 3 4 4 4 3 1 ...
}
\references{
M. Gomtsyan et al. "Variable selection in sparse GLARMA models", arXiv:2007.08623v1
}
\examples{
data(Y)
}
